Refactor movie repository
