+++ src/main/java/com/mithwick93/refactoring/java/Main.java
@@ -3,6 +3,7 @@ package com.mithwick93.refactoring.java;
 import com.mithwick93.refactoring.java.entity.Customer;
 import com.mithwick93.refactoring.java.entity.MovieRental;
 import com.mithwick93.refactoring.java.repositroy.MovieRepository;
+import com.mithwick93.refactoring.java.repositroy.impl.MovieRepositoryImpl;
 
 import java.util.Arrays;
 
@@ -11,7 +12,7 @@ public class Main {
     public static void main(String[] args) {
         String expected = "Rental Record for C. U. Stomer\n\tYou've Got Mail\t3.5\n\tMatrix\t2.0\nAmount owed is 5.5\nYou earned 2 frequent points\n";
 
-        MovieRepository movieRepository = new MovieRepository();
+        MovieRepository movieRepository = new MovieRepositoryImpl();
         String result = new RentalInfo(movieRepository).statement(new Customer("C. U. Stomer", Arrays.asList(new MovieRental("F001", 3), new MovieRental("F002", 1))));
 
         if (!result.equals(expected)) {
