+++ src/main/java/com/mithwick93/refactoring/java/repositroy/MovieRepository.java
@@ -1,22 +1,13 @@
 package com.mithwick93.refactoring.java.repositroy;
 
 import com.mithwick93.refactoring.java.entity.Movie;
-import com.mithwick93.refactoring.java.entity.MovieCode;
 
 import java.util.Map;
 
 /**
- * MovieRepository class to represent a repository of movies
+ * MovieRepository interface to represent a repository of movies
  */
-public class MovieRepository {
-    private final Map<String, Movie> movies;
-
-    /**
-     * Constructor to initialize the repository with some movies
-     */
-    public MovieRepository() {
-        this.movies = initializeMovies();
-    }
+public interface MovieRepository {
 
     /**
      * Get a movie by its id
@@ -24,31 +15,12 @@ public class MovieRepository {
      * @param movieId id of the movie
      * @return Movie with the given id
      */
-    public Movie getMovie(String movieId) {
-        return movies.get(movieId);
-    }
+    Movie getMovie(String movieId);
 
     /**
      * Get all movies
      *
      * @return Map of all movies
      */
-    public Map<String, Movie> getMovies() {
-        return movies;
-    }
-
-    /**
-     * Initialize the repository with some movies
-     *
-     * @return Map of movies
-     */
-    private Map<String, Movie> initializeMovies() {
-        return Map.of(
-                "F001", new Movie("You've Got Mail", MovieCode.REGULAR),
-                "F002", new Movie("Matrix", MovieCode.REGULAR),
-                "F003", new Movie("Cars", MovieCode.CHILDREN),
-                "F004", new Movie("Fast & Furious X", MovieCode.NEW_RELEASE)
-        );
-    }
-
+    Map<String, Movie> getMovies();
 }
